move test button files to more readable location
